Timer improvment

+++ libusb/drivers/controllers/stm32/usbh_api.c
@@ -41,7 +41,7 @@ void USBHstopAllChannels() {
   /* Fifos must be flushed before USBHhaltChannel. */
   FlushTxFifo(ALL_TX_FIFOS);
   FlushRxFifo(ALL_RX_FIFOS);
-  for (i = 0; i < CHNNL_MAX_COUNT; ++i) {
+  for (i = 0; i < (int)CHNNL_MAX_COUNT; ++i) {
     HaltChannel(i);
     P_USB_OTG_HCHNNLS[i].HCINTMSKx = 0;
     P_USB_OTG_HCHNNLS[i].HCINTx = 0xffffffff;
